# lplates/openapi.yaml
openapi: 3.0.3
info:
  title: lplates API
  description: >
    REST API for the lplates driving-lesson booking platform. Covers user
    authentication, instructor management, lesson bookings, EDT (Essential
    Driver Training) progress tracking, instructor pricing, and feedback.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid role supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username or email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Auth]
      summary: Exchange credentials for a JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /lessons/:
    get:
      tags: [Lessons]
      summary: Get all lessons
      responses:
        '200':
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lesson'
        '204':
          description: No lessons found
        '401':
          $ref: '#/components/responses/Unauthorized'
  /lessons/{lessonId}:
    get:
      tags: [Lessons]
      summary: Get a single lesson by ID
      parameters:
        - $ref: '#/components/parameters/LessonId'
      responses:
        '200':
          description: Lesson found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /lessons/create:
    post:
      tags: [Lessons]
      summary: Book a new lesson (LEARNER, ADMIN)
      description: "`studentId` should be the authenticated user's `id`."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lesson'
      responses:
        '200':
          description: Lesson created successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Lesson created successfully with ID: 12"
        '400':
          description: Invalid lessonType
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          description: Failed to create lesson
  /lessons/update/{lessonId}:
    put:
      tags: [Lessons]
      summary: Full update of a lesson
      parameters:
        - $ref: '#/components/parameters/LessonId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lesson'
      responses:
        '200':
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '400':
          description: Invalid lessonType
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Update failed
  /lessons/confirm/{lessonId}:
    post:
      tags: [Lessons]
      summary: Set a lesson's status to confirmed
      parameters:
        - $ref: '#/components/parameters/LessonId'
      responses:
        '200':
          description: Confirmed lesson
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Confirmation failed
  /lessons/pending:
    get:
      tags: [Lessons]
      summary: Get all lessons with status pending
      responses:
        '200':
          description: List of pending lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lesson'
        '204':
          description: No pending lessons found
        '401':
          $ref: '#/components/responses/Unauthorized'
  /lessons/instructor/{instructorId}:
    get:
      tags: [Lessons]
      summary: Get all lessons for an instructor
      parameters:
        - $ref: '#/components/parameters/InstructorId'
      responses:
        '200':
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lesson'
        '204':
          description: No lessons found
        '401':
          $ref: '#/components/responses/Unauthorized'
  /lessons/student/{studentId}:
    get:
      tags: [Lessons]
      summary: Get all lessons for a student
      description: Use the authenticated user's `id` as `studentId`.
      parameters:
        - $ref: '#/components/parameters/StudentId'
      responses:
        '200':
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lesson'
        '204':
          description: No lessons found
        '401':
          $ref: '#/components/responses/Unauthorized'
  /edt/:
    get:
      tags: [EDT Progress]
      summary: Get all EDT progress records (all students)
      responses:
        '200':
          description: List of EDT progress records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EdtProgress'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /edt/{id}:
    get:
      tags: [EDT Progress]
      summary: Get a single EDT progress record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: EDT progress record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdtProgress'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /edt/create:
    post:
      tags: [EDT Progress]
      summary: Create or upsert an EDT progress record (LEARNER, ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EdtProgress'
      responses:
        '201':
          description: Created EDT progress record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdtProgress'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /edt/student/{studentId}:
    get:
      tags: [EDT Progress]
      summary: Get all 12 module records for a student, ordered by moduleNumber
      parameters:
        - $ref: '#/components/parameters/StudentId'
      responses:
        '200':
          description: List of EDT progress records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EdtProgress'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /edt/student/{studentId}/module/{moduleNumber}:
    get:
      tags: [EDT Progress]
      summary: Get a single module record for a student
      parameters:
        - $ref: '#/components/parameters/StudentId'
        - name: moduleNumber
          in: path
          required: true
          description: EDT module number, 1-12
          schema:
            type: integer
            minimum: 1
            maximum: 12
      responses:
        '200':
          description: EDT progress record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdtProgress'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /edt/complete/{studentId}/{moduleNumber}/{lessonId}:
    put:
      tags: [EDT Progress]
      summary: Mark a module as completed and link it to the completing lesson
      parameters:
        - $ref: '#/components/parameters/StudentId'
        - name: moduleNumber
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
        - $ref: '#/components/parameters/LessonId'
      responses:
        '200':
          description: Updated EDT progress record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdtProgress'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    InstructorId:
      name: instructorId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    LessonId:
      name: lessonId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    StudentId:
      name: studentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not permitted to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Invalid username or password"
    Role:
      type: string
      enum: [LEARNER, INSTRUCTOR, ADMIN]
    LessonType:
      type: string
      enum: [lesson, edt, test_car_hire]
...
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: john_doe
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john@example.ie
        password:
          type: string
          format: password
          example: secret123
        role:
          $ref: '#/components/schemas/Role'
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: john_doe
        password:
          type: string
          format: password
          example: secret123
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiJ9..."
        type:
          type: string
          example: Bearer
        userId:
          type: integer
          format: int64
          example: 5
        username:
          type: string
          example: john_doe
        role:
          $ref: '#/components/schemas/Role'
...
    Lesson:
      type: object
      properties:
        lessonId:
          type: integer
          format: int64
        instructorId:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        scheduledDate:
          type: string
          format: date
          example: "2026-08-10"
        scheduledTime:
          type: string
          example: "09:00:00"
        durationMinutes:
          type: integer
          example: 60
        status:
          type: string
          example: pending
        paymentStatus:
          type: string
          example: unpaid
        price:
          type: number
          format: double
          example: 60.00
        currency:
          type: string
          example: EUR
        lessonType:
          $ref: '#/components/schemas/LessonType'
        edtModule:
          type: string
        edtCompleted:
          type: boolean
        notes:
          type: string
        cancellationReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EdtProgress:
      type: object
      properties:
        id:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        moduleNumber:
          type: integer
          minimum: 1
          maximum: 12
        moduleName:
          type: string
        completed:
          type: boolean
        lessonId:
          type: integer
          format: int64
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
...